requirements.txt
autopep8==1.4.3
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
Django==2.1.3
django-cors-headers==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
openapi-codec==1.3.2
psycopg2==2.7.5
pycodestyle==2.4.0
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
uplink==0.5.5
uritemplate==3.0.0
urllib3==1.23
